+++ src/services/clock.rs
@@ -24,7 +24,7 @@ pub fn spawn(service_id: &str, settings: &Settings, tx: &Sender<Message>) -> Res
     let interval = Duration::from_millis(settings.interval_ms);
     let tx = tx.clone();
 
-    supervisor::spawn(format!("my-iot::clock:{}", service_id), move || {
+    supervisor::spawn(format!("my-iot::clock::{}", service_id), move || {
         let mut counter = 1;
         loop {
             tx.send(Message {